+++ src/main/java/br/com/stagiun/tccstagiun/model/service/EmpresaService.java
@@ -5,5 +5,5 @@ import br.com.stagiun.tccstagiun.model.domain.Empresa;
 import java.util.Optional;
 
 public interface EmpresaService extends CrudService<Empresa, Long> {
-    Optional<Empresa> findByCnpj(Integer cnpj);
+    Optional<Empresa> findByCnpj(Long cnpj);
 }
